requirements.txt
aim==3.17.4
numpy==1.23.5
pandas==2.0.1
PyYAML==6.0
scipy==1.10.1
torch==2.0.1
tqdm==4.65.0